feat(v8): Remove `extractTraceparentData` export
