+++ packages/browser/src/index.ts
@@ -69,8 +69,6 @@ export {
   addTracingExtensions,
   setMeasurement,
   // eslint-disable-next-line deprecation/deprecation
-  extractTraceparentData,
-  // eslint-disable-next-line deprecation/deprecation
   getActiveTransaction,
   getSpanStatusFromHttpCode,
   setHttpStatus,
