+++ packages/bun/src/index.ts
@@ -32,8 +32,6 @@ export {
   captureMessage,
   close,
   createTransport,
-  // eslint-disable-next-line deprecation/deprecation
-  extractTraceparentData,
   flush,
   // eslint-disable-next-line deprecation/deprecation
   getActiveTransaction,
