+++ packages/node-experimental/src/index.ts
@@ -64,8 +64,6 @@ export {
   createGetModuleFromFilename,
   close,
   createTransport,
-  // eslint-disable-next-line deprecation/deprecation
-  extractTraceparentData,
   flush,
   Hub,
   runWithAsyncContext,
