+++ packages/remix/src/index.server.ts
@@ -23,8 +23,6 @@ export {
   captureMessage,
   createTransport,
   // eslint-disable-next-line deprecation/deprecation
-  extractTraceparentData,
-  // eslint-disable-next-line deprecation/deprecation
   getActiveTransaction,
   getHubFromCarrier,
   // eslint-disable-next-line deprecation/deprecation
