+++ packages/serverless/src/index.ts
@@ -82,8 +82,6 @@ export {
   getModuleFromFilename,
   createGetModuleFromFilename,
   metrics,
-  // eslint-disable-next-line deprecation/deprecation
-  extractTraceparentData,
   runWithAsyncContext,
   consoleIntegration,
   onUncaughtExceptionIntegration,
