+++ packages/sveltekit/src/server/index.ts
@@ -16,8 +16,6 @@ export {
   withMonitor,
   createTransport,
   // eslint-disable-next-line deprecation/deprecation
-  extractTraceparentData,
-  // eslint-disable-next-line deprecation/deprecation
   getActiveTransaction,
   getHubFromCarrier,
   // eslint-disable-next-line deprecation/deprecation
