+++ packages/tracing-internal/src/exports/index.ts
@@ -1,7 +1,5 @@
 export {
   // eslint-disable-next-line deprecation/deprecation
-  extractTraceparentData,
-  // eslint-disable-next-line deprecation/deprecation
   getActiveTransaction,
   hasTracingEnabled,
   IdleTransaction,
