+++ packages/vercel-edge/src/index.ts
@@ -31,8 +31,6 @@ export {
   captureMessage,
   close,
   createTransport,
-  // eslint-disable-next-line deprecation/deprecation
-  extractTraceparentData,
   flush,
   // eslint-disable-next-line deprecation/deprecation
   getActiveTransaction,
